+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/IDBUtils.kt
@@ -265,7 +265,10 @@ interface IDBUtils {
         title: String,
         desc: String,
         relativePath: String,
-        orientation: Int?
+        orientation: Int?,
+        latitude: Double?,
+        longitude: Double?,
+        creationDate: Long?
     ): AssetEntity {
         var inputStream = ByteArrayInputStream(bytes)
         fun refreshStream() {
@@ -287,7 +290,11 @@ interface IDBUtils {
         )
         val (rotationDegrees, latLong) = Pair(
             orientation ?: if (isAboveAndroidQ) exif.rotationDegrees else 0,
-            if (isAboveAndroidQ) null else exif.latLong
+            if (isAboveAndroidQ) {
+                if (latitude != null && longitude != null) {
+                    doubleArrayOf(latitude, longitude)
+                } else null
+            } else exif.latLong
         )
         refreshStream()
 
@@ -306,7 +313,7 @@ interface IDBUtils {
             put(WIDTH, width)
             put(HEIGHT, height)
             if (isAboveAndroidQ) {
-                put(DATE_TAKEN, timestamp * 1000)
+                put(DATE_TAKEN, creationDate ?: (timestamp * 1000))
                 put(ORIENTATION, rotationDegrees)
                 if (relativePath.isNotBlank()) {
                     put(RELATIVE_PATH, relativePath)
@@ -332,7 +339,10 @@ interface IDBUtils {
         title: String,
         desc: String,
         relativePath: String,
-        orientation: Int?
+        orientation: Int?,
+        latitude: Double?,
+        longitude: Double?,
+        creationDate: Long?
     ): AssetEntity {
         filePath.checkDirs()
         val file = File(filePath)
@@ -357,7 +367,11 @@ interface IDBUtils {
         )
         val (rotationDegrees, latLong) = Pair(
             orientation ?: if (isAboveAndroidQ) exif.rotationDegrees else 0,
-            if (isAboveAndroidQ) null else exif.latLong
+            if (isAboveAndroidQ) {
+                if (latitude != null && longitude != null) {
+                    doubleArrayOf(latitude, longitude)
+                } else null
+            } else exif.latLong
         )
         refreshStream()
 
@@ -381,7 +395,7 @@ interface IDBUtils {
             put(WIDTH, width)
             put(HEIGHT, height)
             if (isAboveAndroidQ) {
-                put(DATE_TAKEN, timestamp * 1000)
+                put(DATE_TAKEN, creationDate ?: (timestamp * 1000))
                 put(ORIENTATION, rotationDegrees)
                 if (relativePath.isNotBlank()) {
                     put(RELATIVE_PATH, relativePath)
@@ -411,7 +425,10 @@ interface IDBUtils {
         title: String,
         desc: String,
         relativePath: String,
-        orientation: Int?
+        orientation: Int?,
+        latitude: Double?,
+        longitude: Double?,
+        creationDate: Long?
     ): AssetEntity {
         filePath.checkDirs()
         val file = File(filePath)
@@ -434,7 +451,11 @@ interface IDBUtils {
         val (rotationDegrees, latLong) = ExifInterface(inputStream).let { exif ->
             Pair(
                 orientation ?: if (isAboveAndroidQ) exif.rotationDegrees else 0,
-                if (isAboveAndroidQ) null else exif.latLong
+                if (isAboveAndroidQ) {
+                    if (latitude != null && longitude != null) {
+                        doubleArrayOf(latitude, longitude)
+                    } else null
+                } else exif.latLong
             )
         }
         refreshStream()
@@ -460,7 +481,7 @@ interface IDBUtils {
             put(WIDTH, info.width)
             put(HEIGHT, info.height)
             if (isAboveAndroidQ) {
-                put(DATE_TAKEN, timestamp * 1000)
+                put(DATE_TAKEN, creationDate ?: (timestamp * 1000))
                 put(ORIENTATION, rotationDegrees)
                 if (relativePath.isNotBlank()) {
                     put(RELATIVE_PATH, relativePath)