+++ ios/Classes/core/PMManager.m
@@ -1,3 +1,4 @@
+#import "NSObject+SafeCheck.h"
 #import "NSString+PM_COMMON.h"
 #import "PHAsset+PM_COMMON.h"
 #import "PHAssetCollection+PM_COMMON.h"
@@ -1270,6 +1271,9 @@
 - (void)saveImage:(NSData *)data
          filename:(NSString *)filename
              desc:(NSString *)desc
+         latitude:(NSNumber *)latitude
+        longitude:(NSNumber *)longitude
+     creationDate:(NSNumber *)creationDate
             block:(AssetBlockResult)block {
     [PMLogUtils.sharedInstance info:[NSString stringWithFormat:@"Saving image with data, length: %lu, filename: %@, desc: %@", (unsigned long)data.length, filename, desc]];
 
@@ -1281,6 +1285,19 @@
         PHAssetResourceCreationOptions *options = [PHAssetResourceCreationOptions new];
         [options setOriginalFilename:filename];
         [request addResourceWithType:PHAssetResourceTypePhoto data:data options:options];
+        
+        // Set location if provided
+        if (![latitude isNilOrNull] && ![longitude isNilOrNull]) {
+            CLLocation *location = [[CLLocation alloc] initWithLatitude:[latitude doubleValue] longitude:[longitude doubleValue]];
+            [request setLocation:location];
+        }
+        
+        // Set creation date if provided
+        if (![creationDate isNilOrNull]) {
+            NSDate *date = [NSDate dateWithTimeIntervalSince1970:[creationDate doubleValue] / 1000.0];
+            [request setCreationDate:date];
+        }
+        
         assetId = request.placeholderForCreatedAsset.localIdentifier;
     }
      completionHandler:^(BOOL success, NSError *error) {
@@ -1301,6 +1318,9 @@
 - (void)saveImageWithPath:(NSString *)path
                  filename:(NSString *)filename
                      desc:(NSString *)desc
+                 latitude:(NSNumber *)latitude
+                longitude:(NSNumber *)longitude
+             creationDate:(NSNumber *)creationDate
                     block:(AssetBlockResult)block {
     NSFileManager *manager = [NSFileManager defaultManager];
     if (![manager fileExistsAtPath:path]) {
@@ -1321,6 +1341,19 @@
             [options setOriginalFilename:filename];
         }
         [request addResourceWithType:PHAssetResourceTypePhoto fileURL:fileURL options:options];
+        
+        // Set location if provided
+        if (![latitude isNilOrNull] && ![longitude isNilOrNull]) {
+            CLLocation *location = [[CLLocation alloc] initWithLatitude:[latitude doubleValue] longitude:[longitude doubleValue]];
+            [request setLocation:location];
+        }
+        
+        // Set creation date if provided
+        if (![creationDate isNilOrNull]) {
+            NSDate *date = [NSDate dateWithTimeIntervalSince1970:[creationDate doubleValue] / 1000.0];
+            [request setCreationDate:date];
+        }
+        
         assetId = request.placeholderForCreatedAsset.localIdentifier;
     }
      completionHandler:^(BOOL success, NSError *error) {
@@ -1341,6 +1374,9 @@
 - (void)saveVideo:(NSString *)path
          filename:(NSString *)filename
              desc:(NSString *)desc
+         latitude:(NSNumber *)latitude
+        longitude:(NSNumber *)longitude
+     creationDate:(NSNumber *)creationDate
             block:(AssetBlockResult)block {
     NSFileManager *manager = [NSFileManager defaultManager];
     if (![manager fileExistsAtPath:path]) {
@@ -1361,6 +1397,19 @@
             [options setOriginalFilename:filename];
         }
         [request addResourceWithType:PHAssetResourceTypeVideo fileURL:fileURL options:options];
+        
+        // Set location if provided
+        if (![latitude isNilOrNull] && ![longitude isNilOrNull]) {
+            CLLocation *location = [[CLLocation alloc] initWithLatitude:[latitude doubleValue] longitude:[longitude doubleValue]];
+            [request setLocation:location];
+        }
+        
+        // Set creation date if provided
+        if (![creationDate isNilOrNull]) {
+            NSDate *date = [NSDate dateWithTimeIntervalSince1970:[creationDate doubleValue] / 1000.0];
+            [request setCreationDate:date];
+        }
+        
         assetId = request.placeholderForCreatedAsset.localIdentifier;
     }
      completionHandler:^(BOOL success, NSError *error) {